+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.3.0 (2 Sep 2016)
 
 * install IDE image configs from https://github.com/ai-traders/ide.git
-* change base image to alpine, so that gitide is now <30MB
+* change base image to alpine, so that gitide is now <33MB
 * do not install Chef-client for tests
 * replace InnerRakefile and Rakefile with only Rakefile